Adds template for posts and create/edit post forms
